Add CharacterPreset class and change Character class.
I had to remove the constructor from the Character class, as the PlayerCharacter class can't use that constructor as a base. As such, I chose to make CharacterPreset a separate class from Character, and implement a constructor in that instead.
All examples of monsters and enemies will use the CharacterPreset class now instead.